requirements.txt
# Recognize - Object and Facial Recognition Application

Flask==3.0.0
Werkzeug==3.0.1
Flask-CORS==4.0.0
python-dotenv==1.0.0

# Computer Vision Libraries
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
numpy==1.26.2

# Deep Learning (Optional - for advanced models)
# tensorflow==2.15.0
# torch==2.1.0
# torchvision==0.16.0

# Image Processing
Pillow==10.1.0

# API and Web
requests==2.31.0
gunicorn==21.2.0

# Task Queue and Message Broker
celery==5.3.4
redis==5.0.1
flower==2.0.1

# Logging and Monitoring
python-json-logger==2.0.7

# Development Dependencies
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.0
flake8==6.1.0
pylint==3.0.3
pre-commit==3.6.0
isort==5.13.2
mypy==1.8.0
bandit==1.7.6
interrogate==1.5.0
safety==2.3.5
